Adds crs
crs works like a cursor for a gap buffer. It only moves the gap on a
change to the buffer and should work rather well for implementing text/bytecode
editors (in my case I'm thinking about a simple "embedded" editor for
methods in a virtual machine). The code is really simple and not super
ultra optimized, but extending it should be a breeze!